removed old tables and fixed bug
- Removed old TicketTable and AssetTable
- Fixed bug where import for delete prompt css was at wrong place
