+++ front/src/App.css
@@ -1,6 +1,9 @@
 @import url("https://fonts.googleapis.com/css2?family=Lato&display=swap");
 
 :root {
+  --tablet-size: 615px;
+  --desktop-size: 1280px;
+
   --background-lighter: #ebeae6;
   --background-darker: #faf4d4;
