--- front/src/components/table/AssetTable.module.css
@@ -1,116 +1,0 @@
-.selected {
-  background-color: var(--pastel-green-dark);
-  transform: translateY(2px);
-  box-shadow: none;
-}
-
-.container {
-  padding-bottom: 5rem;
-}
-
-.container a {
-  color: var(--main-gray);
-  font-weight: normal;
-}
-
-.container .hide-zero {
-  cursor: pointer;
-}
-
-.container input {
-  margin-bottom: 1rem;
-  margin-right: 0.5rem;
-}
-
-.container > div {
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  justify-items: center;
-  margin-bottom: 1rem;
-  position: relative;
-  left: -8%;
-  width: 116%;
-}
-
-.container > div button {
-  border: none;
-  border-radius: 10px;
-  width: 90%;
-  min-height: 1.5rem;
-  cursor: pointer;
-  color: #fff;
-  font-weight: normal;
-  font-size: small;
-  text-transform: capitalize;
-  box-shadow: 0px 3px 5px rgba(0, 0, 0, 0.2);
-  padding: 2px 4px;
-  background-color: var(--pastel-green);
-}
-
-.container > div button:hover {
-  background-color: var(--pastel-green-dark);
-}
-
-.container > div button:active {
-  transform: translateY(2px);
-  box-shadow: none;
-}
-
-.container > div button.selected {
-  background-color: var(--pastel-green-dark);
-  transform: translateY(2px);
-  box-shadow: none;
-}
-
-.container .container-ticker-name {
-  text-transform: uppercase;
-}
-
-.red {
-  color: #ea526f;
-}
-
-.green {
-  color: #57cc99;
-}
-
-.container .table {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  left: -8%;
-  width: 116%;
-}
-
-.container .table .thead .tr {
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  justify-items: center;
-  text-transform: uppercase;
-  font-weight: bold;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.2);
-  padding-bottom: 0.4rem;
-  margin-bottom: 0.5rem;
-}
-
-.container .table .tbody .tr {
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  justify-items: center;
-  align-items: center;
-  height: 2rem;
-  margin-bottom: 0.5rem;
-  border-radius: 0.5rem;
-  border: 1.5px solid rgba(0, 0, 0, 0.2);
-  box-shadow: 3px 3px 5px rgba(0, 0, 0, 0.2);
-  padding: 0 0.3rem;
-}
-
-.table .tbody a:nth-child(odd) .tr {
-  background-color: rgba(255, 255, 255, 0.5);
-}
-
-.table .tbody a:nth-child(even) .tr {
-  background-color: rgba(252, 234, 222, 0.5);
-}
