+++ front/src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
+import "react-confirm-alert/src/react-confirm-alert.css";
 import { BrowserRouter as Router } from "react-router-dom";
 import { AuthContextProvider } from "./context/auth-context";
 import { WalletContextProvider } from "./context/wallet-context";
