+++ front/src/pages/overview/Overview.module.css
@@ -10,10 +10,6 @@
   border-bottom: 1px solid rgba(0, 0, 0, 0.2);
 }
 
-.section__assets {
-  margin: 0 auto;
-}
-
 .pie-chart__pie-chart {
   width: var(--pie-dimension);
   height: var(--pie-dimension);
